Add rubric-specific news lists with refresh and retry
